Update static-pods.md
The lines that are modified are some items in an ordered list, and each item has more than one paragrahs. In this scenario, it seems that  the beginning serial number of the item may become uncorrect with only Markdown synax. I tried to solve this by inserting some HTML code. Wonder whether this is appropriate and if there is a better method to tackle this formatting problem. Be grateful to know : )